docs: add launch GIF set + shared capture pipeline
Adds the Phase 2 launch-asset GIFs (wall-sit, jumping-jacks, jab-cross,
dead-bug) in docs/launch-media/, rendered from the real playground.

Extracts the capture core into scripts/lib/gif-capture.mjs (with an
optional per-target vertical crop anchor) so the README pipeline
(capture-gifs.mjs) and the new launch pipeline (capture-launch-gifs.mjs)
share one implementation.
